Same for ampEnvelope
